+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Fix guard-let shorthand syntax.
 - Fix accuracy of unused parameter analysis for overridden and protocol conforming functions.
+- Fix retaining buildFinalResult(_:) and buildLimitedAvailability(_:) result builder methods.
 
 ## 2.10.1 (2022-11-20)